Version 7.1.40 - Fixed spawn percentage issue related to the hunter enchantment

+++ src/main/java/com/magmaguy/elitemobs/mobspawning/NaturalMobSpawnEventHandler.java
@@ -82,7 +82,7 @@ public class NaturalMobSpawnEventHandler implements Listener {
 
         ArrayList<Player> nearbyPlayers = PlayerScanner.getNearbyPlayers(livingEntity.getLocation());
 
-        int huntingGearChanceAdder = HunterEnchantment.getHuntingGearBonus(nearbyPlayers);
+        double huntingGearChanceAdder = HunterEnchantment.getHuntingGearBonus(nearbyPlayers);
         validChance += huntingGearChanceAdder;
 
         if (ValidWorldsConfig.fileConfiguration.getBoolean("Nightmare mode worlds." + event.getEntity().getWorld().getName()))
